requirements/requirements.txt
## dependencias a serem instaladas
nose
behave
selenium
behave-html-formatter
allure-behave
faker
pyautogui